refactor: change file directory structure
