--- week8/algoorgoal/src/common/components/Row.tsx
@@ -1,20 +1,0 @@
-import React from 'react';
-import styled from 'styled-components';
-import { JustifyContentType } from '../types/JustifyContentType';
-
-const StyledRow = styled.span<{
-  justifyContent?: JustifyContentType;
-}>`
-  display: flex;
-  width: 100%;
-  justify-content: ${(props) => props.justifyContent || 'flex-start'};
-`;
-
-interface RowPropsType {
-  children: React.ReactNode;
-  justifyContent?: JustifyContentType;
-}
-
-export default function Row({ children, ...props }: RowPropsType) {
-  return <StyledRow {...props}>{children}</StyledRow>;
-}
