+++ week8/algoorgoal/src/components/Store/Menu/FoodCategoryRow/index.tsx
@@ -1,0 +1,19 @@
+import React from 'react';
+import Row from '../../../common/Row';
+import Text from '../../../common/Text';
+
+interface FoodCategoryRowPropsType {
+  foodCategory: string;
+}
+
+export default function FoodCategoryRow({
+  foodCategory,
+}: FoodCategoryRowPropsType) {
+  return (
+    <Row justifyContent="flex-start">
+      <Text color="muted" size="medium">
+        {foodCategory}
+      </Text>
+    </Row>
+  );
+}
