+++ week8/algoorgoal/src/components/Store/Menu/FoodItem/FoodImage/index.styles.tsx
