+++ week8/algoorgoal/src/components/Store/Menu/FoodItem/FoodName/index.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import Text from '../../../../common/Text';
+
+interface FoodNamePropsType {
+  children: React.ReactNode;
+}
+export default function FoodName({ children, ...props }: FoodNamePropsType) {
+  return (
+    <Text color="highlight" size="medium" {...props}>
+      {children}
+    </Text>
+  );
+}
