+++ week8/algoorgoal/src/components/Store/StoreProfile/StoreImage/index.tsx
