--- week8/algoorgoal/src/components/StoreProfile/StoreDetail/StoreMinimumOrderValue/index.tsx
@@ -1,16 +1,0 @@
-import React from 'react';
-import Text from '../../../../common/components/Text';
-
-interface StoreMinimumOrderValuePropsType {
-  minimumOrderValue: number;
-}
-
-export default function StoreMinimumOrderValue({
-  minimumOrderValue,
-}: StoreMinimumOrderValuePropsType) {
-  return (
-    <Text color="base" size="small">
-      최소주문: {minimumOrderValue}원
-    </Text>
-  );
-}
