+++ week8/algoorgoal/src/pages/Home/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Text from '../../common/components/Text';
+import Text from '../../components/common/Text';
 
 export default function Home() {
   return (
