+++ week8/algoorgoal/src/pages/Order/index.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function Order() {
+  return <div>Order</div>;
+}
