+++ week8/algoorgoal/src/stories/FoodImage.stories.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import type { Meta, StoryObj } from '@storybook/react';
 
-import FoodImage from '../components/Menu/FoodItem/FoodImage';
+import FoodImage from '../components/Store/Menu/FoodItem/FoodImage';
 
 const meta = {
   title: 'week8/FoodImage',
